requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic-settings>=2.1.0

# LangChain / LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0

# HTTP Client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
